FONT tag is inserted after the code is pretty printed into string.
It is really agly implementation. The structure is once completely
destructed by pretty printing the object into a STRING and
re-interpreted in order to insert FONT tags. Some additional
markers (brackets and braces) are put before the pretty printing and
pretty printed as well as the code itself to make it possible to
distinguish which part is should be wrapped by FONT tags.

However, it seems to be difficult to prevent. If the FONT tag is
inserted before the pretty printing, or simultaneous to that, then
pretty print became completely disgusting.

I'm going to check the detail of pretty printing of Common Lisp, and try
to find out a better approach.